-- calc max distance branch --

+++ app/src/main/java/com/FreedAsd/fuel_calculator/domain/max_distance/DistanceResultDomain.kt
@@ -3,7 +3,7 @@ package com.FreedAsd.fuel_calculator.domain.max_distance
 interface DistanceResultDomain {
 
     data class Base(
-        private val maxDistance: Int,
+        private val maxDistance: Float,
         private val tripPrice: Float
     ) : DistanceResultDomain
 }
